+++ src/pageEditor/tabs/editTab/EditTab.module.scss
@@ -106,10 +106,6 @@
   max-width: 360px;
 }
 
-.brickFaIcon {
-  color: inherit;
-}
-
 .tabPane {
   height: 100%;
 }
